requirements.txt
quart
quart-cors
python-dotenv